backend/requirements.txt
flask==3.0.3
flask-cors==4.0.0
python-dotenv==1.0.1
requests==2.32.3
pdfminer.six==20231228
python-docx==1.1.2
beautifulsoup4==4.12.3
lxml==5.2.1
websocket-client==1.8.0
